Add option to put text at bottom of list instead of top The existing 'para' functionality was set to add arbitrary HTML to a list, but only at index 0 - i.e. becoming the first item in the list.
This commit adds an HTML <select> field to choose between adding the text as a 'header' (existing functionality) or a 'footer'. We do this by grabbing the number of items in the list from the first call, and making that the index, since it's indexed at zero.
New variables in the php to enable this:

$POSITION - this is either 'header' or 'footer' and taken from the select field. Defaults to 'header' as per earlier behaviour.
$positionIndex - either equal to meta.item_count from the list (if position is 'footer' otherwise 0 as per earlier behaviour.

The final PATCH call to the API now uses $positionIndex for relationships.meta.index rather than the hard coded 0.